chore: made into class, wrote demonstration code
